common_imports.py
from app.conf import base as settings
from django.utils.translation import ugettext_lazy as _
from rest_framework.views import APIView
from oauth2_provider.contrib.rest_framework import TokenHasReadWriteScope
from django.http import JsonResponse
from rest_framework.parsers import JSONParser, MultiPartParser
from datetime import date, timedelta, datetime


__all__ = ['_', 'APIView', 'settings', 'TokenHasReadWriteScope', 'JsonResponse',
           'JSONParser', 'MultiPartParser',
           'date', 'timedelta', 'datetime',
           ]
