requirements.txt
argon2-cffi==19.1.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-rest-coreapi-schema==0.1.2
django-user-agents==0.3.2
djangorestframework==3.7.7
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.1.1
oauthlib==3.0.1
Pillow==5.4.1
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-memcached==1.59
pytz==2018.9
PyYAML==3.13
requests==2.21.0
shortuuid==0.5.0
six==1.12.0
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.24.1
user-agents==1.1.0
